+++ Tower Defense/Assets/Scripts/Turrets/EnemyTest.cs
@@ -44,7 +44,7 @@ public class EnemyTest : MonoBehaviour, ITurretDamage, ISlowable, IPoisonable, I
         print("I have been slowed for " + secondsSlowed + " seconds with a reduction of " + slowReduction);
     }
 
-    public void Poison(float secondsPoisoned, float poisonDamagePerSecond)
+    public void Poison(float secondsPoisoned, int poisonDamagePerSecond)
     {
         print("I have been poisoned for " + secondsPoisoned + " seconds with a damage of " + poisonDamagePerSecond + " per second");
     }
